refactor: Remove search functionality and cache feature
