--- src/api/search/search.router.js
@@ -1,12 +1,0 @@
-import express from 'express';
-
-import { catchAsyncErrorHandler } from '../../app.middlewares.js';
-import { db, redis } from '../../database/db.js';
-import { SearchService } from './search.service.js';
-import { getSearchHandler } from './search.handler.js';
-
-const search = express.Router();
-
-search.get('/', catchAsyncErrorHandler(getSearchHandler(SearchService(db, redis))));
-
-export { search };
