--- src/api/user/user.service.js
@@ -1,10 +1,0 @@
-export function UserService(db) {
-	return {
-		getUsers: async () => {
-			return await db.select('*').from('users');
-		},
-		getTenantUsers: async (tenantId) => {
-			return await db.select('*').from('users').where({ tenant_id: tenantId });
-		},
-	};
-}
